// src/app/core/models/masters/compliance-management/compliance-type.ts
export interface ComplianceType {
    id: number;
    title: string;
    description: string;
    status_id :number;
    status_label: string;
}
export interface ComplianceTypePaginationResponse {
    current_page: number;
    total: number;
    per_page: number;
    last_page: number;
    from: number;
    data: ComplianceType[];
}